Add additional player gold check in state
In case of a player wanting to build multiple factories in a single
turn, the command giver doesn not maintain a rolling count of the number
factories that need to be produced.

This temporarily adds a check to ensure that a factory is not produced
if the player does not have enough gold. However, this does not send an
explicit player error. It fails silently.

TODO: Fix This